+++ examples/react-router-rsc/vite.config.ts
@@ -2,18 +2,12 @@ import { unstable_reactRouterRSC as reactRouterRSC } from '@react-router/dev/vit
 import rsc from '@vitejs/plugin-rsc';
 import tailwindcss from '@tailwindcss/vite';
 import { defineConfig } from 'vite';
-import tsconfigPaths from 'vite-tsconfig-paths';
 import mdx from 'fumadocs-mdx/vite';
 import * as MdxConfig from './source.config';
 
 export default defineConfig({
-  plugins: [
-    mdx(MdxConfig),
-    tailwindcss(),
-    reactRouterRSC(),
-    rsc(),
-    tsconfigPaths({
-      projects: ['./tsconfig.json'],
-    }),
-  ],
+  plugins: [mdx(MdxConfig), tailwindcss(), reactRouterRSC(), rsc()],
+  resolve: {
+    tsconfigPaths: true,
+  },
 });