+++ packages/radix-ui/tsdown.config.ts
@@ -26,6 +26,6 @@ export default defineConfig({
     console.log('CSS import updated');
   },
   deps: {
-    onlyAllowBundle: [],
+    onlyBundle: [],
   },
 });