+++ packages/stf/tsdown.config.ts
@@ -9,7 +9,7 @@ export default defineConfig({
     sourcemap: false,
   },
   deps: {
-    onlyAllowBundle: [],
+    onlyBundle: [],
   },
   exports: true,
 });